feat(trackpredict): refactor scope hacking from confirmtrack
moved promptForTrack() and underlying implementation from confirmtrack into trackmanager instead.

BREAKING CHANGE: promptForTrack() and launchConfirmTrack() moved to TrackManager.getInstance()
